+++ app/assets/javascripts/fae/form/_ajax.js
@@ -20,8 +20,9 @@ Fae.form.ajax = {
     this.htmlListeners();
     this.applyCookies();
 
-    // Re-applied once AJAX form is loaded in
     this.deleteNoForm();
+
+    // Re-applied once AJAX form is loaded in
     if (this.$filter_form.length) {
       this.filterSelect();
       this.filterSubmission();
@@ -126,7 +127,6 @@ Fae.form.ajax = {
           FCH.smoothScroll($this.find('.js-addedit-form-wrapper'), 500, 100, 120);
         }
 
-        _this.deleteNoForm();
         if (_this.$filter_form.length) {
           _this.filterSelect();
           _this.filterSubmission();
@@ -135,9 +135,11 @@ Fae.form.ajax = {
         Fae.navigation.fadeNotices();
 
       } else if ($target.hasClass('js-asset-delete-link')) {
-        // handle remove asset links
-        $target.parent().fadeOut('fast', function() {
-          $(this).next('.asset-inputs').fadeIn('fast');
+        // handle remove asset links on nested forms
+        var $parent = $target.closest('.asset-actions');
+
+        $parent.fadeOut(function(){
+          $parent.next('.asset-inputs').fadeIn();
         });
       }
     });
@@ -272,12 +274,13 @@ Fae.form.ajax = {
   },
 
   /**
-   * On deletes that don't exist in a form like file upload area
+   * Delete or replace file for non-AJAX'd fields
    */
   deleteNoForm: function() {
     $('.js-asset-delete-link').on('ajax:success', function(){
       var $this = $(this);
       if (!$this.closest('.js-addedit-form-wrapper').length) {
+        // handle remove asset links
         var $parent = $this.closest('.asset-actions');
 
         $parent.fadeOut(function(){