+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - enhancements
 - bugs
     + \#52680: Add spacing between label and helper text on checkbox fields
+    + \#51982: Vertical checkboxes should be vertical
+    + \#52696: Remove image/file on replace image for nested forms too
 
 ## 1.2.1